backend/requirements.txt
pathway[all]
python-dotenv
requests
flask
flask-cors
sentence-transformers
